Moved EVSE component related classes to proper package
